requirements.txt
dj-database-url==0.5.0
django==2.0.6
django-crispy-forms==1.7.2
django-image-tools==0.8b3
django-image-tools-2==1.0
django-markdown-deux==1.0.5
django-multiselectfield==0.1.8
django-pagedown==1.0.6
gunicorn==19.9.0
markdown2==2.3.7
mysqlclient==1.3.12
Pillow==5.0.0
psycopg2==2.7.5
pygame==1.9.5
pytz==2018.5
virtualenv==16.0.0
